+++ src/main/java/com/example/demo/builder/builder/SearchWindowBuilder.java
@@ -11,9 +11,6 @@ import com.example.demo.service.ReservationInfoService;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.util.Optional;
 
 import static javax.swing.WindowConstants.EXIT_ON_CLOSE;
 
@@ -86,6 +83,54 @@ public abstract class SearchWindowBuilder {
         else {
             searchWindow.setStatusInfo(new JLabel("이용가능"));
         }
+    }
+
+    public void buildBookInfoLabelBuilder(){
+        Book searchedBook = searchWindow.getSearchedBook();
+        searchWindow.setTitleInfo(new JLabel(searchedBook.getTitle()));
+        searchWindow.setPositionInfo(new JLabel(searchedBook.getPosition()));
+        buildStatusInfo();
+        searchWindow.setIsbnInfo(new JLabel(String.valueOf(searchedBook.getIsbn())));
+        searchWindow.setPublisherInfo(new JLabel(searchedBook.getPublisher()));
+
+        addBookInfoLabel();
+    }
+
+    public void buildBookInfoInputLabelBuilder() {
+        Book searchedBook = searchWindow.getSearchedBook();
+        searchWindow.setTitleInput(new JTextField(searchedBook.getTitle()));
+        searchWindow.setPositionInput(new JTextField(searchedBook.getPosition()));
+        buildStatusInfo();
+        searchWindow.setIsbnInput(new JTextField(searchedBook.getIsbn().toString()));
+        searchWindow.setPublisherInput(new JTextField(searchedBook.getPublisher()));
+
+        addBookInfoInputField();
+    }
+
+    private void addBookInfoLabel(){
+        searchWindow.add(searchWindow.getTitleLabel());
+        searchWindow.add(searchWindow.getTitleInfo());
+        searchWindow.add(searchWindow.getPosition());
+        searchWindow.add(searchWindow.getPositionInfo());
+        searchWindow.add(searchWindow.getStatus());
+        searchWindow.add(searchWindow.getStatusInfo());
+        searchWindow.add(searchWindow.getIsbn());
+        searchWindow.add(searchWindow.getIsbnInfo());
+        searchWindow.add(searchWindow.getPublisher());
+        searchWindow.add(searchWindow.getPublisherInfo());
+    }
+
+    private void addBookInfoInputField(){
+        searchWindow.add(searchWindow.getTitleLabel());
+        searchWindow.add(searchWindow.getTitleInput());
+        searchWindow.add(searchWindow.getPosition());
+        searchWindow.add(searchWindow.getPositionInput());
+        searchWindow.add(searchWindow.getStatus());
+        searchWindow.add(searchWindow.getStatusInfo());
+        searchWindow.add(searchWindow.getIsbn());
+        searchWindow.add(searchWindow.getIsbnInput());
+        searchWindow.add(searchWindow.getPublisher());
+        searchWindow.add(searchWindow.getPublisherInput());
     }
 
     public abstract void buildBackButton();
