+++ src/main/java/com/example/demo/builder/concreteBuilder/SearchWindowAdminEditBuilder.java
@@ -22,26 +22,6 @@ public class SearchWindowAdminEditBuilder extends SearchWindowBuilder {
 
     @Override
     public void buildBookInfoLabel() {
-        Book searchedBook = searchWindow.getSearchedBook();
-        searchWindow.setTitleInput(new JTextField(searchedBook.getTitle()));
-        searchWindow.setPositionInput(new JTextField(searchedBook.getPosition()));
-        buildStatusInfo();
-        searchWindow.setIsbnInput(new JTextField(searchedBook.getIsbn().toString()));
-        searchWindow.setPublisherInput(new JTextField(searchedBook.getPublisher()));
-
-        addBookInfoInputField();
-    }
-
-    private void addBookInfoInputField(){
-        searchWindow.add(searchWindow.getTitleLabel());
-        searchWindow.add(searchWindow.getTitleInput());
-        searchWindow.add(searchWindow.getPosition());
-        searchWindow.add(searchWindow.getPositionInput());
-        searchWindow.add(searchWindow.getStatus());
-        searchWindow.add(searchWindow.getStatusInfo());
-        searchWindow.add(searchWindow.getIsbn());
-        searchWindow.add(searchWindow.getIsbnInput());
-        searchWindow.add(searchWindow.getPublisher());
-        searchWindow.add(searchWindow.getPublisherInput());
+        super.buildBookInfoInputLabelBuilder();
     }
 }
