+++ src/main/java/com/example/demo/builder/concreteBuilder/SearchWindowUserBuilder.java
@@ -3,13 +3,11 @@ package com.example.demo.builder.concreteBuilder;
 import com.example.demo.actionListener.SearchActionListener;
 import com.example.demo.builder.builder.SearchWindowBuilder;
 import com.example.demo.domain.Book;
-import com.example.demo.jframe.SearchWindow;
-
 
 import javax.swing.*;
 import java.awt.*;
 
-public class SearchWindowAdminBuilder extends SearchWindowBuilder {
+public class SearchWindowUserBuilder extends SearchWindowBuilder {
 
     @Override
     public void buildBackButton() {
@@ -21,16 +19,7 @@ public class SearchWindowAdminBuilder extends SearchWindowBuilder {
 
     @Override
     public void buildFunctionButton() {
-        searchWindow.setEditBTN(new Button("수정하기"));
-        searchWindow.getEditBTN().setBounds(20, 5, 70, 30);
-        searchWindow.getEditBTN().addActionListener(new SearchActionListener());
 
-        searchWindow.setDeleteBTN(new Button("삭제하기"));
-        searchWindow.getDeleteBTN().setBounds(20, 5, 70, 30);
-        searchWindow.getDeleteBTN().addActionListener(new SearchActionListener());
-
-        searchWindow.add(searchWindow.getEditBTN());
-        searchWindow.add(searchWindow.getDeleteBTN());
     }
 
     @Override
